result: deserialize rows lazily
If we look at the CassResult methods, there is no method that requests
a CassRow at given index. It means, that we can lazily deserialize rows
during iteration (`cass_iterator_from_result`).

The only exception is `cass_result_first_row`. The result should
hold first deserialized row, and simply return it. We should not
deserialize the row each time this method is called.

This is exactly what this commit does. It adjusts both `CassRowsResult`
and `CassRowsResultIterator`:
- CassRowsResult now holds only first row, and all deserialization payload (i.e.
  metadata and serialized rows).
- CassRowsResultIterator from now on holds a TypedRowIterator, instead
  of reference to CassRowsResult.

This semantic is more efficient, and is employed by cpp-driver as well.