row: borrow metadata from result, instead of cloning an Arc
Because of that, we need to introduce a lifetime parameter to CassRow.
CassRow will from now on borrow metadata from CassResult (CassRowsResult).

OTOH, notice that CassRowsResult need to hold the deserialized first row.
This is because `cass_result_first_row` method exists. We don't want to
deserialize the row each time this method is called.

Because of that, the self-borrowing appears in our codebase. We have a
CassRowsResult, which holds a CassRow that borrows from the CassRowsResult.
It's not possible to express it in safe Rust, thus we need to erase the lifetime
of this specific CassRow to 'static. It is fine, because we know that
this CassRow won't live longer than CassResult that it borrows metadata from.

The iterators are adjusted as well - we no longer need to hold metadata in Arc
in CassRowsResultIterator.