iterator: make CassResultIterator an enum
It is to distinguish between the two types of iterators that we have:
- iterator over non-rows result (always returning null)
- iterator over rows result

In one of the next commits, we will modify `CassRowsResultIterator`, so
it holds `TypedRowIterator` instead of reference to CassRowsResult.

This change would be necessary then anyway, so it's better to do it now
and reduce the noise during review in later commit.